requirements.txt
# Core dependencies for AIstylist project
Flask==3.0.0
Werkzeug==3.0.0
requests==2.31.0

# OpenAI API dependencies
openai==1.92.2
python-dotenv==1.1.1

# Image processing dependencies (minimal set)
Pillow==11.2.1
numpy==1.26.4

# Timezone handling
pytz==2024.1

# OAuth for Google login
Flask-Dance==7.0.0

# Payment processing
stripe==10.12.0

# Mobile image format support
pillow-heif>=1.1.0

# Scheduled tasks
apscheduler>=3.11.0

# Alternative AI models
together>=1.5.0
Flask-Dance==7.0.0 